feat: #2931 add auto truncation of long tags
