+++ mlflow/utils/autologging_utils/client.py
@@ -28,8 +28,8 @@ from mlflow.utils.validation import (
     MAX_METRICS_PER_BATCH,
     MAX_PARAM_VAL_LENGTH,
     MAX_PARAMS_TAGS_PER_BATCH,
-    MAX_TAG_VAL_LENGTH,
 )
+from mlflow.utils.validation_common import MAX_TAG_VAL_LENGTH
 
 _logger = logging.getLogger(__name__)
