+++ mlflow/utils/validation_common.py
@@ -1,0 +1,1 @@
+MAX_TAG_VAL_LENGTH = 5000
